Avoid using volatile and cache line false sharing
Each notifier occupies a full cache line to avoid false sharing.
Use CPU relax (see cpu_relax() from kernel) instead of volatile for
busy-wating. See
https://www.kernel.org/doc/Documentation/volatile-considered-harmful.txt
for some more details why one should avoid volatile.
Unfortunately the kernel cpu_relax() function is not available to users
and is architecture dependent. Because we do not want to rely on kernel
sources to be installed we have to provide an implementation of this
function for each architecture. This commit only adds support for x86.