Implement communication using send/receive
Adds support for hardware that is not capable of RDMA Writes with
immediate data. The immediate data is required to inform the remote side
of RDMA event.
By default send/receive is used, add definition VERBS_RDMA to enable the
old RDMA Write behavior. This should be done specifically for each
cluster, hence no user interaction is required.
So far, if supported, RDMA has shown to be the better choice.