Add warning for self and null provider
These two providers might by used accidientially and result in hardly
debuggable errors or at least result in bad performance in case of self.
The self provider is there just for compatibility, hence it is not highly
optimized. Indeed it can not be further optiomized in the framework of
pMR (any further optimization would required threaded memory copy).