Allow fine-grained scoped locking by (optional) mutex parameter
So far all scopedLocks have been using a global Mutex. This is useful to
lock access to e.g. MPI if required. I.e. access to many functions, of
which only one might be called at a time. But this might have a
performance impact for cases where one certain function is not allowed
to be called more than once at a time, but hase no dependencies on any
other function. For this the new optional parameter can be used to
specify a Function specific mutex.