Support RMA without Event; Replace context -> ID
Add support for RMA without RMA Event capability by adding CQ data
(similar to tag for tagged messaging) to message and poll for receive of
such message. Therefor context has been replaced by ID, as it can either
be a context or a CQ data. In both cases it is a unique (to connection)
message identification.