Return of the SCIF
Promotes the auxiliary provider SCIF to a fully functional provider.
This feature has already been in the old C version, but so far has never
been added in the newer framework as it seemed to be of no interested.
With signs of a potential x200 coprocessor release, this might change in
future. Anyway, is also useful for KNC systems without any IB hardware.