Set defaults for QPACE3
Set OFI defaults to suit Omni-Path/psm2 and allow to co-coexist with
OpenMPI 1.10.3 (mtl ofi). This will not work with OpenMPI set to use
psm2 directly. Due to OpenMPI not asking OFI for RMA (and other)
capabilities, we are limited to tagged messages. Furthermore enable
shared memory data transfer (CMA). The CMA implementation seems to
perform better on QPACE3 than the psm2 built-in shared memory mechanism.