Add experimental PSM2 multi-thread and wrapper
The multi-threaded optimization pretend psm2 to be thread safe.
According to the documentation it is not. However the OFI psm2 provider
claims to be almost thread safe without any heavy locking for the MQ.
Thus lets simply assume the same.
The libwrapper *should* allow to directly work with the PSM2 MQ,
although MPI (or any other application) has already initialized PSM2.
Currently this only works for applications that use PSM2 through OFI.
That is due to all other known applications (OpenMPI and MVAPICH) always
assume the context of a retrieved status to be a valid pointer to one of
their own internal data. Within libfabric (luckily) a NULL context is
simply ignored.

Both features are marked as experimental - use at your own risk!